Got the path unit tests working again, found a bunch of bugs in the settings code, fixed those bugs.  Almost certainly includes a fix for HER-1275, but that needs more testing.
* .classpath
Include new common/src/test/resources for test data.

* commons/src/test/java/org/archive/settings/path/Foo.java
* commons/src/test/java/org/archive/settings/path/Bar.java
* commons/src/test/java/org/archive/settings/path/Baz.java
New test module classes.

* commons/src/test/java/org/archive/settings/path/PathTestBase.java
* commons/src/test/java/org/archive/settings/path/PathValidatorTest.java
* commons/src/test/java/org/archive/settings/path/PathListerTest.java
Use Foo/Bar/Baz instead of extractors that no longer live in this project.

* commons/src/test/java/org/archive/settings/path/PathChangerTest.java
Use PathChanger to clone the PathTestBase sheets; much more coverage that way.

* commons/src/main/java/org/archive/state/Path.java
(equals)
(hashCode): So these can be used as hashtable keys.

* commons/src/main/java/org/archive/state/KeyTypes.java
(<clinit>): Make bigdecimal tag map to BigDecimal, not BigInteger

* commons/src/main/java/org/archive/settings/SheetManager.java
(getGlobalSheet): No longer abstract, as there's only one correct impl

* commons/src/main/java/org/archive/settings/MemorySheetManager.java
(<clinit>): Register keys with KeyManager.
(globals): Deleted; committing global sheet never replaced this field
(root): Deleted; committing global sheet never replaced this field
(<init>): Automatically create ROOT map, also set value for MANAGER
(commit): Remember to announceChanges()

* commons/src/main/java/org/archive/settings/SingleSheet.java
(resolveMap)
(resolveList): Changed the way sheets are reported on these.  If the map/list
is defined both in the global sheet and in an override sheet, then the 
reported sheet list now contains "global,override" instead of just "global".
(resolveEditableList)
(resolveEditableMap): If the current sheet doesn't actually contain an 
override for the map or list, return null.  Otherwise the global sheet gets
modified instead of an override sheet.

* commons/src/main/java/org/archive/settings/MultiTypedMap.java
(entrySet): Don't use a LinkedHashSet of Map.Entry, since the 
Map.Entry.getValue() instances will never be equal, resulting in multiple
entries in the set for the same key.  Instead use a LinkedHashMap and take
its entrySet.

* commons/src/main/java/org/archive/settings/path/PathValidator.java
(ex): Fixed typo that made reported problems hard to read.
(validateKey): Better exception message when no such key exists.
(resolve): Cleared some compiler warnings.

* commons/src/main/java/org/archive/settings/path/PathLister.java
(consume): Look at first element instead of last per changes in map/list
sheet resolution reporting up in SingleSheet.

* commons/src/main/java/org/archive/settings/path/PathChanger.java
(finish): Use validate instead of check for the common case; needed due to
SingleSheet.resolveEditableMap/List changes.  For list resolution, adjust the
index in the actual path by the number of elements in the global sheet if 
appropriate.

+++ heritrix2/commons/src/main/java/org/archive/state/KeyTypes.java
@@ -112,7 +112,7 @@ public class KeyTypes {
         types.put(String.class, "string");
         types.put(Pattern.class, "pattern");
         types.put(BigInteger.class, "biginteger");
-        types.put(BigInteger.class, "bigdecimal");
+        types.put(BigDecimal.class, "bigdecimal");
         types.put(Enum.class, ENUM_TAG);
         types.put(Path.class, "file"); 
         TYPES = Collections.unmodifiableMap(types);
